improve the vertical layout algorithm
The previous algorithm didn't consider that balancing
the dimensions of the grid can both yield additional capacity
and get a good cell aspect ratio. For example, if you need to
place 4 items on a small vertical screen, the algorithm will
first add three rows, and no columns.

The issue is that at this point, there are no good options
without balancing dimensions:
- adding a column brings the capacity up to 6, which is
  too much as it makes the last row empty
- adding another row doesn't yield a good cell aspect ratio

In this case, removing a row and adding a column in one
operation gets the algorithm to the optimal solution.